Moved UG REST Client to separate package improved tests
