+++ apis/my-studies-api/__test__/callDummyUser.spec.ts
@@ -1,5 +1,5 @@
 import { describe, expect, jest, test } from '@jest/globals';
-import { UGRestClient } from "../src/ugRestClient";
+import { UGRestClient } from "kpm-ug-rest-client";
 import { TUgUser } from "../src/api";
 import { addMockResponse } from "../__mocks__/openid-client"
