+++ apis/my-studies-api/evolene-local-packages.json
@@ -1,1 +1,1 @@
-{"kpm-api-common":{"path":"packages/kpm-api-common"}}
+{"kpm-api-common":{"path":"packages/kpm-api-common"},"kpm-ug-rest-client":{"path":"packages/kpm-ug-rest-client"}}
