+++ apis/my-studies-api/package.json
@@ -24,6 +24,7 @@
     "dotenv": "^16.0.2",
     "express": "^4.18.1",
     "kpm-api-common": "^0.1.0",
+    "kpm-ug-rest-client": "^0.1.0",
     "openid-client": "^5.1.9",
     "ts-node": "^10.9.1"
   }
