+++ packages/kpm-ug-rest-client/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "kpm-ug-rest-client",
+  "version": "0.1.0",
+  "description": "KPM UG REST API Client",
+  "main": "src/index.ts",
+  "scripts": {
+    "test": "jest",
+    "release": ""
+  },
+  "author": "KTH Royal Institute of Technology",
+  "license": "MIT",
+  "devDependencies": {
+    "@types/express": "^4.17.13",
+    "prettier": "^2.7.1",
+    "ts-jest": "^29.0.1",
+    "typescript": "^4.8.3"
+  },
+  "dependencies": {
+    "openid-client": "^5.1.10"
+  }
+}
