+++ packages/kpm-ug-rest-client/src/index.ts
@@ -1,0 +1,1 @@
+export * from "./ugRestClient";
